fix(app): deal with layout issues
Although 'height: 100%' works, it is inconsistent and
leads to issues with storybook 'padded' layout.
When set, 'height: 100%' forces the browser to always consider content
to be up to window height, even if it overflows.
This is probably the reason it is inconsistent.
Those implications can be seen with DevTools open and 'Inspect Element' pointed to body/#root.
 The box model (browser hint of how it 'sees' the element,
  position, and it's size) does not match with content size.