fix(app): do not mix view and logic for login page
LoginPage should contain important logic while
LoginView contains visual look and a simple logic (form logic).
LoginForm component is removed because it mostly duplicates logic for a small readability gain.
End-to-end test is removed because there is no backend server configured,
plus it should only validate important paths.
Testing loading indicator is not that important and it should be moved to component tests.