fix(app): use same query key in single vehicle view or all vehicles view
Because the same function for querying is used in both places
this change results in small optimization for the end user.
While you are viewing single vehicle for a long time (> 5 min), when you go back to all vehicles,
there is no need to refetch again and the page renders instantly (no loading indicator rendered).